+++ react/components/EditorContainer/Sidebar/ComponentEditor/Form.tsx
@@ -8,6 +8,7 @@ import Dropdown from '../../../form/Dropdown'
 import ErrorListTemplate from '../../../form/ErrorListTemplate'
 import FieldTemplate from '../../../form/FieldTemplate'
 import ImageUploader from '../../../form/ImageUploader'
+import IOMessage from '../../../form/IOMessage'
 import ObjectFieldTemplate from '../../../form/ObjectFieldTemplate'
 import Radio from '../../../form/Radio'
 import TextArea from '../../../form/TextArea'
@@ -16,6 +17,7 @@ import Toggle from '../../../form/Toggle'
 export const widgets = {
   BaseInput: BaseInput as Widget,
   CheckboxWidget: Toggle as Widget,
+  IOMessage: IOMessage as Widget,
   RadioWidget: Radio as Widget,
   SelectWidget: Dropdown as Widget,
   TextareaWidget: TextArea as Widget,