+++ react/components/EditorContainer/Sidebar/ConfigurationList/LayoutEditor/index.tsx
@@ -1,5 +1,4 @@
 import React, { Component } from 'react'
-import { injectIntl } from 'react-intl'
 import { IChangeEvent } from 'react-jsonschema-form'
 
 import { getBlockPath, getRelativeBlocksIds } from '../../../../../utils/blocks'
@@ -11,7 +10,7 @@ import { UpdateBlockMutationFn } from '../../../mutations/UpdateBlock'
 import ComponentEditor from '../../ComponentEditor'
 import { FormMetaContext, ModalContext } from '../../typings'
 
-interface Props extends ReactIntl.InjectedIntlProps {
+interface Props {
   editor: EditorContext
   formMeta: FormMetaContext
   iframeRuntime: RenderContext
@@ -36,12 +35,11 @@ class LayoutEditor extends Component<Props> {
   }
 
   public render() {
-    const { editor, formMeta, iframeRuntime, modal } = this.props
+    const { formMeta, iframeRuntime, modal } = this.props
 
     return (
       <ComponentEditor
         data={this.getExtensionProps()}
-        editor={editor}
         iframeRuntime={iframeRuntime}
         isLoading={formMeta.isLoading && !modal.isOpen}
         onChange={this.handleChange}
@@ -68,18 +66,17 @@ class LayoutEditor extends Component<Props> {
   }
 
   private handleChange = (event: IChangeEvent) => {
-    const {
-      editor: { editTreePath },
-      formMeta,
-      intl,
-      iframeRuntime,
-    } = this.props
+    const { editor, formMeta, iframeRuntime } = this.props
 
     if (!formMeta.wasModified) {
       formMeta.setWasModified(true)
     }
 
-    updateExtensionFromForm(editTreePath, event, intl, iframeRuntime)
+    updateExtensionFromForm({
+      data: event,
+      runtime: iframeRuntime,
+      treePath: editor.editTreePath!,
+    })
   }
 
   private handleDiscard = () => {
@@ -157,4 +154,4 @@ class LayoutEditor extends Component<Props> {
   }
 }
 
-export default injectIntl(LayoutEditor)
+export default LayoutEditor