+++ react/components/EditorContainer/Sidebar/index.tsx
@@ -4,11 +4,10 @@ import { defineMessages, injectIntl } from 'react-intl'
 import Modal from '../../Modal'
 
 import Content from './Content'
-import { FormMetaConsumer } from './FormMetaContext'
-import { ModalConsumer } from './ModalContext'
+import { useFormMetaContext } from './FormMetaContext'
+import { useModalContext } from './ModalContext'
 
 interface CustomProps {
-  editor: EditorContext
   highlightHandler: (treePath: string | null) => void
   runtime: RenderContext
 }
@@ -31,46 +30,41 @@ const messages = defineMessages({
 })
 
 const Sidebar: React.FunctionComponent<Props> = ({
-  editor,
   highlightHandler,
   intl,
   runtime,
-}) => (
-  <div
-    id="sidebar-vtex-editor"
-    className="z-1 h-100 top-3em-ns calc--height-ns w-18em-ns w-100 w-auto-ns flex flex-row-reverse overflow-x-auto"
-  >
-    <FormMetaConsumer>
-      {formMeta => (
-        <ModalConsumer>
-          {modal => (
-            <nav
-              id="admin-sidebar"
-              className="transition animated fadeIn b--light-silver bw1 z-2 h-100 pt8 pt0-ns calc--height-ns overflow-x-hidden w-100 font-display bg-white shadow-solid-x w-18em-ns admin-sidebar"
-            >
-              <div className="h-100 flex flex-column dark-gray">
-                <Modal
-                  isActionLoading={formMeta.isLoading}
-                  isOpen={modal.isOpen}
-                  onClickAction={modal.actionHandler}
-                  onClickCancel={modal.cancelHandler}
-                  onClose={modal.close}
-                  textButtonAction={intl.formatMessage(messages.save)}
-                  textButtonCancel={intl.formatMessage(messages.discard)}
-                  textMessage={intl.formatMessage(messages.unsaved)}
-                />
-                <Content
-                  editor={editor}
-                  highlightHandler={highlightHandler}
-                  iframeRuntime={runtime}
-                />
-              </div>
-            </nav>
-          )}
-        </ModalConsumer>
-      )}
-    </FormMetaConsumer>
-  </div>
-)
+}) => {
+  const formMeta = useFormMetaContext()
+  const modal = useModalContext()
+
+  return (
+    <div
+      id="sidebar-vtex-editor"
+      className="z-1 h-100 top-3em-ns calc--height-ns w-18em-ns w-100 w-auto-ns flex flex-row-reverse overflow-x-auto"
+    >
+      <nav
+        id="admin-sidebar"
+        className="transition animated fadeIn b--light-silver bw1 z-2 h-100 pt8 pt0-ns calc--height-ns overflow-x-hidden w-100 font-display bg-white shadow-solid-x w-18em-ns admin-sidebar"
+      >
+        <div className="h-100 flex flex-column dark-gray">
+          <Modal
+            isActionLoading={formMeta.isLoading}
+            isOpen={modal.isOpen}
+            onClickAction={modal.actionHandler}
+            onClickCancel={modal.cancelHandler}
+            onClose={modal.close}
+            textButtonAction={intl.formatMessage(messages.save)}
+            textButtonCancel={intl.formatMessage(messages.discard)}
+            textMessage={intl.formatMessage(messages.unsaved)}
+          />
+          <Content
+            highlightHandler={highlightHandler}
+            iframeRuntime={runtime}
+          />
+        </div>
+      </nav>
+    </div>
+  )
+}
 
 export default injectIntl(Sidebar)